+++ packages/ai/src/generate-text/content-part.ts
@@ -2,6 +2,7 @@ import { ProviderMetadata } from '../types';
 import { Source } from '../types/language-model';
 import { GeneratedFile } from './generated-file';
 import { ToolApprovalRequestOutput } from './tool-approval-request-output';
+import { ReasoningOutput } from './reasoning-output';
 import { TypedToolCall } from './tool-call';
 import { TypedToolError } from './tool-error';
 import { TypedToolResult } from './tool-result';
@@ -10,7 +11,7 @@ import { ToolSet } from './tool-set';
 // TODO AI SDK 5.1 / AI SDK 6: revisit naming, e.g. rename to Output
 export type ContentPart<TOOLS extends ToolSet> =
   | { type: 'text'; text: string; providerMetadata?: ProviderMetadata }
-  | { type: 'reasoning'; text: string; providerMetadata?: ProviderMetadata }
+  | ReasoningOutput
   | ({ type: 'source' } & Source)
   | { type: 'file'; file: GeneratedFile; providerMetadata?: ProviderMetadata } // different because of GeneratedFile object
   | ({ type: 'tool-call' } & TypedToolCall<TOOLS> & {