+++ packages/cohere/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/cohere
 
+## 2.0.10
+
+### Patch Changes
+
+- Updated dependencies [0294b58]
+  - @ai-sdk/provider-utils@3.0.9
+
 ## 2.0.9
 
 ### Patch Changes