+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,33 @@
 
 ### Patch Changes
 
+- ai@5.0.44
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [0294b58]
+  - ai@5.0.43
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [de5c066]
+  - ai@5.0.42
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [cd91e4b]
+  - ai@5.0.41
+
+## 0.0.1
+
+### Patch Changes
+
 - ai@5.0.40
 
 ## 0.0.1